sample update : merge viewmodel
