+++ src/Lemon.ModuleNavigation.SampleViewModel/BaseNavigationViewModel.cs
@@ -3,7 +3,7 @@ using ReactiveUI;
 using System.Diagnostics;
 using System.Reactive;
 
-namespace Lemon.ModuleNavigation.WpfSample;
+namespace Lemon.ModuleNavigation.SampleViewModel;
 
 public class BaseNavigationViewModel : ReactiveObject, INavigationAware
 {
