+++ src/Lemon.ModuleNavigation.SampleViewModel/ViewBetaViewModel.cs
@@ -3,22 +3,18 @@ using Lemon.ModuleNavigation.Core;
 using ReactiveUI;
 using System.Reactive;
 
-namespace Lemon.ModuleNavigation.WpfSample.ViewModels;
+namespace Lemon.ModuleNavigation.SampleViewModel;
 
-public class ViewAlphaViewModel : BaseNavigationViewModel, IDialogAware
+public class ViewBetaViewModel : BaseNavigationViewModel, IDialogAware
 {
-    public ViewAlphaViewModel()
-    {
-    }
-
-    public string Title => nameof(ViewAlphaViewModel);
+    public string Title => nameof(ViewBetaViewModel);
 
     public event Action<IDialogResult>? RequestClose;
-    public ReactiveCommand<Unit, Unit> CloseCommand => ReactiveCommand.Create(() =>
+    public ReactiveCommand<Unit, Unit> CloseCommand => ReactiveCommand.Create(() => 
     {
         var param = new DialogParameters
         {
-            { "from", nameof(ViewAlphaViewModel) }
+            { "from", nameof(ViewBetaViewModel) }
         };
         RequestClose?.Invoke(new DialogResult(ButtonResult.OK, param));
     });
@@ -35,14 +31,13 @@ public class ViewAlphaViewModel : BaseNavigationViewModel, IDialogAware
 
     public void OnDialogClosed()
     {
-       
+        
     }
 
     public void OnDialogOpened(IDialogParameters? parameters)
     {
         IsDialog = true;
     }
-
     public override bool IsNavigationTarget(NavigationContext navigationContext)
     {
         if (navigationContext.Parameters is not null)
