+++ src/Lemon.ModuleNavigation.WpfSample/Lemon.ModuleNavigation.WpfSample.csproj
@@ -15,6 +15,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Lemon.ModuleNavigation.SampleViewModel\Lemon.ModuleNavigation.SampleViewModel.csproj" />
     <ProjectReference Include="..\Lemon.ModuleNavigation.Wpf\Lemon.ModuleNavigation.Wpf.csproj" />
   </ItemGroup>
