+++ src/Lemon.ModuleNavigation.sln
@@ -24,6 +24,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lemon.ModuleNavigation.Wpf"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lemon.ModuleNavigation.WpfSample", "Lemon.ModuleNavigation.WpfSample\Lemon.ModuleNavigation.WpfSample.csproj", "{8EC545BD-E076-A2DF-A7A4-2F11221AF297}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lemon.ModuleNavigation.SampleViewModel", "Lemon.ModuleNavigation.SampleViewModel\Lemon.ModuleNavigation.SampleViewModel.csproj", "{99600F69-5901-4CEA-8C54-BFDC0751E207}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -62,6 +64,10 @@ Global
 		{8EC545BD-E076-A2DF-A7A4-2F11221AF297}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8EC545BD-E076-A2DF-A7A4-2F11221AF297}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8EC545BD-E076-A2DF-A7A4-2F11221AF297}.Release|Any CPU.Build.0 = Release|Any CPU
+		{99600F69-5901-4CEA-8C54-BFDC0751E207}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{99600F69-5901-4CEA-8C54-BFDC0751E207}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{99600F69-5901-4CEA-8C54-BFDC0751E207}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{99600F69-5901-4CEA-8C54-BFDC0751E207}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
